+++ src/main/scala/eu/sim642/adventofcodelib/Geometry.scala
@@ -9,7 +9,7 @@ object Geometry {
    * Calculates the area of a simple polygon using the shoelace formula.
    * @see [[https://en.wikipedia.org/wiki/Shoelace_formula]]
    */
-  def polygonArea(poss: collection.Seq[Pos]): Int = {
+  def polygonArea(poss: collection.Seq[Pos]): Long = { // TODO: generalize return type
     ((poss.last +: poss).iterator
       .zipWithTail
       .map(_ cross _)
